+++ tests/testthat/test-get_microdata.R
@@ -23,17 +23,17 @@ test_that("download and read data", {
   # testthat::expect_error(get_marco(folder = 1, toR = FALSE))
 })
 
-context("Testing get_dictionary")
-test_that("download dictionaries", {
-  testthat::expect_error(get_dictionary(year = 2010, folder = tempdir()))
-  get_dictionary(year = 2019, folder = tempdir())
-  archivo <- fs::dir_ls(tempdir(), regexp = "\\.xls$")
-  archivo <- archivo[which.max(file.info(archivo)$mtime)]
-  testthat::expect_equal(length(archivo), 1)
-  get_dictionary(year = NULL, folder = tempdir())
-  archivo <- fs::dir_ls(tempdir(), regexp = "\\.xls$")
-  archivo <- archivo[which.max(file.info(archivo)$mtime)]
-  testthat::expect_equal(length(archivo), 1)
-  testthat::expect_error(get_dictionary(folder = 1))
-  testthat::expect_message(get_dictionary(year = 2019, folder = tempdir()))
-})
+# context("Testing get_dictionary")
+# test_that("download dictionaries", {
+#   testthat::expect_error(get_dictionary(year = 2010, folder = tempdir()))
+#   get_dictionary(year = 2019, folder = tempdir())
+#   archivo <- fs::dir_ls(tempdir(), regexp = "\\.xls$")
+#   archivo <- archivo[which.max(file.info(archivo)$mtime)]
+#   testthat::expect_equal(length(archivo), 1)
+#   get_dictionary(year = NULL, folder = tempdir())
+#   archivo <- fs::dir_ls(tempdir(), regexp = "\\.xls$")
+#   archivo <- archivo[which.max(file.info(archivo)$mtime)]
+#   testthat::expect_equal(length(archivo), 1)
+#   testthat::expect_error(get_dictionary(folder = 1))
+#   testthat::expect_message(get_dictionary(year = 2019, folder = tempdir()))
+# })